requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
